remove oc_adjustment_row wrapper
